Refactor cards directory and upgrade dev dependencies
