+++ core/src/main/resources/org/jobrunr/dashboard/frontend/src/components/overview/server-analytics/estimated-processing-time-card.js
@@ -1,6 +1,6 @@
 import {useJobStats} from "../../../hooks/useJobStats";
 import {SuffixFreeTimeAgo} from "../../utils/time-ago";
-import {AnalyticsCard} from "./analytics-card.js";
+import {AnalyticsCard} from "../../ui/AnalyticsCard.js";
 import {AccessTime} from "@mui/icons-material";
 
 const EstimatedProcessingTimeCard = () => {
