+++ core/src/main/resources/org/jobrunr/dashboard/frontend/src/components/overview/server-analytics/realtime-graph.js
