+++ core/src/main/resources/org/jobrunr/dashboard/frontend/src/components/ui/AnalyticsCard.js
